TAP-103: Initial style scrubbing
