+++ packages/article/article-header.js
@@ -9,7 +9,7 @@ import {
 } from "@times-components/article-flag";
 import ArticleLabel from "@times-components/article-label";
 import { Text, View } from "react-native";
-import styles from "./styles/article-header-style";
+import styles from "./styles/header";
 
 const flagsMapping = new Map([
   ["NEW", <NewArticleFlag />],
