--- packages/article/styles/article-header-style/index.js
@@ -1,14 +1,0 @@
-import { StyleSheet, Platform } from "react-native";
-import globalStyle from "../article-global-style";
-import sharedStyles from "./shared";
-import nativeStyles from "./native";
-
-const styles = {};
-
-if (Platform.OS !== "web") {
-  Object.assign(styles, globalStyle, StyleSheet.create(nativeStyles));
-} else {
-  Object.assign(styles, globalStyle, StyleSheet.create(sharedStyles));
-}
-
-export default styles;
