--- packages/article/styles/article-style/native.js
@@ -1,10 +1,0 @@
-import sharedStyles from "./shared";
-
-const nativeStyles = {
-  ...sharedStyles,
-  articleTextElement: Object.assign({}, sharedStyles.articleTextElement, {
-    marginBottom: 20
-  })
-};
-
-export default nativeStyles;
