--- packages/article/styles/article-style/web.js
@@ -1,10 +1,0 @@
-import sharedStyles from "./shared";
-
-const webStyles = {
-  ...sharedStyles,
-  articleTextElement: Object.assign({}, sharedStyles.articleTextElement, {
-    marginTop: 0
-  })
-};
-
-export default webStyles;
