+++ packages/article/styles/body/index.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+import sharedStyles from './shared';
+import globalStyle from "../article-global-style";
+
+const webStyles = {
+  ...sharedStyles,
+  articleTextElement: {
+    ...sharedStyles.articleTextElement,
+    marginTop: 0
+  }
+};
+
+const styles = StyleSheet.create({
+  ...globalStyle,
+  ...webStyles
+});
+
+export default styles;
