+++ packages/article/styles/body/shared.js
@@ -1,17 +1,13 @@
-import constants from "../const";
-
-const { fontFamilyBody, colourText } = constants;
-
 const sharedStyles = {
   leadAsset: {
     marginBottom: 10
   },
   articleTextElement: {
-    fontFamily: fontFamilyBody,
+    fontFamily: "TimesDigitalW04",
     lineHeight: 26,
     fontSize: 17,
     marginBottom: 25,
-    color: colourText
+    color: "#333333"
   }
 };
