--- packages/article/styles/const/index.js
@@ -1,7 +1,0 @@
-export default {
-  fontFamilyBody: "TimesDigitalW04",
-  fontFamilyStandFirst: "TimesModern-Regular",
-  colourText: "#333333",
-  colourHeadline: "#000000",
-  colourBorder: "#D0CECE"
-};
